+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/CoreModule.java
@@ -37,5 +37,7 @@ public interface CoreModule extends ModuleDescriptor {
 	String TOPIC_DELEGATION_DELETED_TO_DELEGATOR = String.format("%s:delegationDeletedToDelegator", MODULE_ID);
 	String TOPIC_DELEGATION_INSTANCE_CREATED_TO_DELEGATE = String.format("%s:delegationInstanceCreatedToDelegate", MODULE_ID);
 	String TOPIC_TWO_FACTOR_VERIFICATION_CODE = String.format("%s:twoFactorVerificationCode", MODULE_ID);
+	String TOPIC_APPROVE_ROLE_DEFINITION_CHANGE = String.format("%s:approveRoleDefinitionChange", MODULE_ID);
+	String TOPIC_DISAPPROVE_ROLE_DEFINITION_CHANGE = String.format("%s:disapproveRoleDefinitionChange", MODULE_ID);
 	
 }