+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [11.0.0]
+
+### Administrator
+
+- 🟡 [#2684](https://redmine.czechidm.com/issues/2684) - **Some messages are missing in the Logging events.** 
+  - **Why:** This fix solves an issue where some logged messages cannot be saved **to the database** (because the message is too long or contains prohibited characters). The solution is to use a DB appender that solves this. 
+  - **What to do:** Change a log database appender from ``ch.qos.logback.classic.db.DBAppender`` to ``eu.bcvsolutions.idm.core.exception.IdmDBAppender`` in the log configuration file (**lockback-spring.xml**).
+
 ## [10.8.0]
 
 ### Administrator